Move AI outputs to shared
