+++ src/components/notebook/cell/ExecutableCell.tsx
@@ -32,7 +32,7 @@ import { SqlToolbar } from "./toolbars/SqlToolbar.js";
 
 import { MaybeCellOutputs } from "@/components/outputs/MaybeCellOutputs.js";
 import { useToolApprovals } from "@/hooks/useToolApprovals.js";
-import { AiToolApprovalOutput } from "../../outputs/AiToolApprovalOutput.js";
+import { AiToolApprovalOutput } from "../../outputs/shared-with-iframe/AiToolApprovalOutput.js";
 
 // Cell-specific styling configuration
 const getCellStyling = (cellType: "code" | "sql" | "ai") => {
