+++ src/components/outputs/index.ts
@@ -1,9 +1,9 @@
 // Re-export AnsiOutput from the notebook folder for convenience
 import { AiToolCallData, AiToolResultData } from "@/schema";
 
-export { AiToolCallOutput } from "./AiToolCallOutput.js";
-export { AiToolResultOutput } from "./AiToolResultOutput.js";
-export { AiToolApprovalOutput } from "./AiToolApprovalOutput.js";
+export { AiToolCallOutput } from "./shared-with-iframe/AiToolCallOutput.js";
+export { AiToolResultOutput } from "./shared-with-iframe/AiToolResultOutput.js";
+export { AiToolApprovalOutput } from "./shared-with-iframe/AiToolApprovalOutput.js";
 export { HtmlOutput } from "./shared-with-iframe/HtmlOutput.js";
 export { ImageOutput } from "./shared-with-iframe/ImageOutput.js";
 export { JsonOutput } from "./shared-with-iframe/JsonOutput.js";
