+++ src/components/outputs/shared-with-iframe/AiToolApprovalOutput.tsx
@@ -1,6 +1,6 @@
 import React from "react";
-import { Button } from "../ui/button";
-import { Card } from "../ui/card";
+import { Button } from "../../ui/button";
+import { Card } from "../../ui/card";
 import {
   Shield,
   ShieldCheck,
