+++ src/components/outputs/shared-with-iframe/AiToolResultOutput.tsx
